requirements.txt
Django==2.2.16
pytz==2018.5
django-filter==23.2
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
drf-yasg==1.21.7
psycopg2-binary==2.9.6
python-dotenv==1.0.0
python3-openid==3.2.0
PyYAML==6.0